Added files via upload
This is  version 2. It is in beta. I have ran some test and it look ok but use with care :)
It enables the inclusion of numpy.1darray input. if you want to use scalar, you have to specify the index corresponding to the scalar e.g x0=autoD.Scalar(index=0),  x0=autoD.Scalar(index=1).